requirements.txt
streamlit>=1.24.0
Pillow>=10.0.0
groq>=0.3.0
httpx>=0.24.0
openai>=1.0.0
pyperclip>=1.8.2
toml>=0.10.2
transformers>=4.30.0
torch>=2.0.0
timm>=0.9.0
python-dotenv>=1.0.0
logging>=0.5.1